# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DDBJ Search API",
    "description": "RESTful API for searching and retrieving BioProject, BioSample, SRA, and JGA entries from DDBJ.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/entries/{type}/bulk": {
      "post": {
        "tags": [
          "Bulk"
        ],
        "summary": "Bulk entry retrieval",
        "description": "Retrieve multiple entries by their IDs.  Up to 1000 IDs can be specified per request.\n\n**format=json** (default): returns ``{entries: [...], notFound: [...]}``.\n\n**format=ndjson**: returns one entry per line in NDJSON format (Content-Type: application/x-ndjson).  IDs not found are not included in the NDJSON output.",
        "operationId": "bulk_entries_entries__type__bulk_post",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/DbType",
              "description": "Database type."
            },
            "description": "Database type."
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/BulkFormat",
              "description": "Response format: 'json' (JSON Array) or 'ndjson' (Newline Delimited JSON).",
              "default": "json"
            },
            "description": "Response format: 'json' (JSON Array) or 'ndjson' (Newline Delimited JSON)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (e.g. deep paging limit exceeded).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found (entry does not exist or invalid type).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity (parameter validation error).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Attribute": {
        "properties": {
          "attribute_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attribute Name"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "harmonized_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Harmonized Name"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "attribute_name",
          "display_name",
          "harmonized_name",
          "content"
        ],
        "title": "Attribute"
      },
      "BioProject": {
        "properties": {
          "identifier": {
            "type": "string",
            "title": "Identifier"
          },
          "properties": {
            "title": "Properties"
          },
          "distribution": {
            "items": {
              "$ref": "#/components/schemas/Distribution"
            },
            "type": "array",
            "title": "Distribution"
          },
          "isPartOf": {
            "type": "string",
            "const": "BioProject",
            "title": "Ispartof"
          },
          "type": {
            "type": "string",
            "const": "bioproject",
            "title": "Type"
          },
          "objectType": {
            "type": "string",
            "enum": [
              "UmbrellaBioProject",
              "BioProject"
            ],
            "title": "Objecttype"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "organism": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Organism"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "organization": {
            "items": {
              "$ref": "#/components/schemas/Organization"
            },
            "type": "array",
            "title": "Organization"
          },
          "publication": {
            "items": {
              "$ref": "#/components/schemas/Publication"
            },
            "type": "array",
            "title": "Publication"
          },
          "grant": {
            "items": {
              "$ref": "#/components/schemas/Grant"
            },
            "type": "array",
            "title": "Grant"
          },
          "externalLink": {
            "items": {
              "$ref": "#/components/schemas/ExternalLink"
            },
            "type": "array",
            "title": "Externallink"
          },
          "dbXrefs": {
            "items": {
              "$ref": "#/components/schemas/Xref"
            },
            "type": "array",
            "title": "Dbxrefs"
          },
          "sameAs": {
            "items": {
              "$ref": "#/components/schemas/Xref"
            },
            "type": "array",
            "title": "Sameas"
          },
          "status": {
            "type": "string",
            "enum": [
              "live",
              "unpublished",
              "suppressed",
              "withdrawn"
            ],
            "title": "Status"
          },
          "accessibility": {
            "type": "string",
            "enum": [
              "public-access",
              "controlled-access"
            ],
            "title": "Accessibility"
          },
          "dateCreated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datecreated"
          },
          "dateModified": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datemodified"
          },
          "datePublished": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datepublished"
          }
        },
        "type": "object",
        "required": [
          "identifier",
          "properties",
          "distribution",
          "isPartOf",
          "type",
          "objectType",
          "name",
          "url",
          "organism",
          "title",
          "description",
          "organization",
          "publication",
          "grant",
          "externalLink",
          "dbXrefs",
          "sameAs",
          "status",
          "accessibility",
          "dateCreated",
          "dateModified",
          "datePublished"
        ],
        "title": "BioProject"
      },
...
      "BioSample": {
        "properties": {
          "identifier": {
            "type": "string",
            "title": "Identifier"
          },
          "properties": {
            "title": "Properties"
          },
          "distribution": {
            "items": {
              "$ref": "#/components/schemas/Distribution"
            },
            "type": "array",
            "title": "Distribution"
          },
          "isPartOf": {
            "type": "string",
            "const": "BioSample",
            "title": "Ispartof"
          },
          "type": {
            "type": "string",
            "const": "biosample",
            "title": "Type"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "organism": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Organism"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "attributes": {
            "items": {
              "$ref": "#/components/schemas/Attribute"
            },
            "type": "array",
            "title": "Attributes"
          },
          "model": {
            "items": {
              "$ref": "#/components/schemas/Model"
            },
            "type": "array",
            "title": "Model"
          },
          "package": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Package"
              },
              {
                "type": "null"
              }
            ]
          },
          "dbXrefs": {
            "items": {
              "$ref": "#/components/schemas/Xref"
            },
            "type": "array",
            "title": "Dbxrefs"
          },
          "sameAs": {
            "items": {
              "$ref": "#/components/schemas/Xref"
            },
            "type": "array",
            "title": "Sameas"
          },
          "status": {
            "type": "string",
            "enum": [
              "live",
              "unpublished",
              "suppressed",
              "withdrawn"
            ],
            "title": "Status"
          },
          "accessibility": {
            "type": "string",
            "enum": [
              "public-access",
              "controlled-access"
            ],
            "title": "Accessibility"
          },
          "dateCreated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datecreated"
          },
          "dateModified": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datemodified"
          },
          "datePublished": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datepublished"
          }
        },
        "type": "object",
        "required": [
          "identifier",
          "properties",
          "distribution",
          "isPartOf",
          "type",
          "name",
          "url",
          "organism",
          "title",
          "description",
          "attributes",
          "model",
          "package",
          "dbXrefs",
          "sameAs",
          "status",
          "accessibility",
          "dateCreated",
          "dateModified",
          "datePublished"
        ],
        "title": "BioSample"
      },
...
      "BulkFormat": {
        "type": "string",
        "enum": [
          "json",
          "ndjson"
        ],
        "title": "BulkFormat",
        "description": "Output format for the Bulk API."
      },
      "BulkRequest": {
        "properties": {
          "ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Ids",
            "description": "List of entry identifiers to retrieve (max 1000)."
          }
        },
        "type": "object",
        "required": [
          "ids"
        ],
        "title": "BulkRequest",
        "description": "Request body for POST /entries/{type}/bulk."
      },
      "BulkResponse": {
        "properties": {
          "entries": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BioProject"
                },
                {
                  "$ref": "#/components/schemas/BioSample"
                },
                {
                  "$ref": "#/components/schemas/SRA"
                },
                {
                  "$ref": "#/components/schemas/JGA"
                }
              ]
            },
            "type": "array",
            "title": "Entries",
            "description": "Found entries (raw ES documents)."
          },
          "notFound": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notfound",
            "description": "IDs that were not found."
          }
        },
        "type": "object",
        "required": [
          "entries",
          "notFound"
        ],
        "title": "BulkResponse",
        "description": "Response for POST /entries/{type}/bulk (format=json).\n\n``entries`` contains the found entries as raw ES documents.\n``notFound`` lists IDs that could not be found."
      },
      "DbType": {
        "type": "string",
        "enum": [
          "bioproject",
          "biosample",
          "sra-submission",
          "sra-study",
          "sra-experiment",
          "sra-run",
          "sra-sample",
          "sra-analysis",
          "jga-study",
          "jga-dataset",
          "jga-dac",
          "jga-policy"
        ],
        "title": "DbType",
        "description": "Database types supported by the API (12 types)."
      },
...
      "Distribution": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "encodingFormat": {
            "type": "string",
            "title": "Encodingformat"
          },
          "contentUrl": {
            "type": "string",
            "title": "Contenturl"
          }
        },
        "type": "object",
        "required": [
          "type",
          "encodingFormat",
          "contentUrl"
        ],
        "title": "Distribution"
      },
...
      "ExternalLink": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "url",
          "label"
        ],
        "title": "ExternalLink"
      },
...
      "Grant": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "agency": {
            "items": {
              "$ref": "#/components/schemas/Agency"
            },
            "type": "array",
            "title": "Agency"
          }
        },
        "type": "object",
        "required": [
          "title",
          "agency"
        ],
        "title": "Grant"
      },
      "JGA": {
        "properties": {
          "identifier": {
            "type": "string",
            "title": "Identifier"
          },
          "properties": {
            "title": "Properties"
          },
          "distribution": {
            "items": {
              "$ref": "#/components/schemas/Distribution"
            },
            "type": "array",
            "title": "Distribution"
          },
          "isPartOf": {
            "type": "string",
            "const": "jga",
            "title": "Ispartof"
          },
          "type": {
            "type": "string",
            "enum": [
              "jga-study",
              "jga-dataset",
              "jga-dac",
              "jga-policy"
            ],
            "title": "Type"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "organism": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Organism"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "dbXrefs": {
            "items": {
              "$ref": "#/components/schemas/Xref"
            },
            "type": "array",
            "title": "Dbxrefs"
          },
          "sameAs": {
            "items": {
              "$ref": "#/components/schemas/Xref"
            },
            "type": "array",
            "title": "Sameas"
          },
          "status": {
            "type": "string",
            "const": "live",
            "title": "Status"
          },
          "accessibility": {
            "type": "string",
            "const": "controlled-access",
            "title": "Accessibility"
          },
          "dateCreated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datecreated"
          },
          "dateModified": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datemodified"
          },
          "datePublished": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datepublished"
          }
        },
        "type": "object",
        "required": [
          "identifier",
          "properties",
          "distribution",
          "isPartOf",
          "type",
          "name",
          "url",
          "organism",
          "title",
          "description",
          "dbXrefs",
          "sameAs",
          "status",
          "accessibility",
          "dateCreated",
          "dateModified",
          "datePublished"
        ],
        "title": "JGA"
      },
...
      "Model": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "Model"
      },
      "Organism": {
        "properties": {
          "identifier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Identifier"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "identifier",
          "name"
        ],
        "title": "Organism"
      },
      "Organization": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "organizationType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organizationtype"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "abbreviation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abbreviation"
          }
        },
        "type": "object",
        "required": [
          "name",
          "organizationType",
          "role",
          "url",
          "abbreviation"
        ],
        "title": "Organization"
      },
...
      "ProblemDetails": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Problem type URI.",
            "default": "about:blank"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Short human-readable summary."
          },
          "status": {
            "type": "integer",
            "title": "Status",
            "description": "HTTP status code."
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Human-readable explanation."
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance",
            "description": "Request path where the error occurred."
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp",
            "description": "Error timestamp (ISO 8601)."
          },
          "requestId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestid",
            "description": "Request tracking ID (same as X-Request-ID header)."
          }
        },
        "type": "object",
        "required": [
          "title",
          "status",
          "detail"
        ],
        "title": "ProblemDetails",
        "description": "RFC 7807 Problem Details error response."
      },
      "Publication": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "Reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "DbType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dbtype"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "title",
          "date",
          "Reference",
          "url",
          "DbType",
          "status"
        ],
        "title": "Publication"
      },
      "SRA": {
        "properties": {
          "identifier": {
            "type": "string",
            "title": "Identifier"
          },
          "properties": {
            "title": "Properties"
          },
          "distribution": {
            "items": {
              "$ref": "#/components/schemas/Distribution"
            },
            "type": "array",
            "title": "Distribution"
          },
          "isPartOf": {
            "type": "string",
            "const": "sra",
            "title": "Ispartof"
          },
          "type": {
            "type": "string",
            "enum": [
              "sra-submission",
              "sra-study",
              "sra-experiment",
              "sra-run",
              "sra-sample",
              "sra-analysis"
            ],
            "title": "Type"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "organism": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Organism"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "dbXrefs": {
            "items": {
              "$ref": "#/components/schemas/Xref"
            },
            "type": "array",
            "title": "Dbxrefs"
          },
          "sameAs": {
            "items": {
              "$ref": "#/components/schemas/Xref"
            },
            "type": "array",
            "title": "Sameas"
          },
          "status": {
            "type": "string",
            "enum": [
              "live",
              "unpublished",
              "suppressed",
              "withdrawn"
            ],
            "title": "Status"
          },
          "accessibility": {
            "type": "string",
            "enum": [
              "public-access",
              "controlled-access"
            ],
            "title": "Accessibility"
          },
          "dateCreated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datecreated"
          },
          "dateModified": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datemodified"
          },
          "datePublished": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Datepublished"
          }
        },
        "type": "object",
        "required": [
          "identifier",
          "properties",
          "distribution",
          "isPartOf",
          "type",
          "name",
          "url",
          "organism",
          "title",
          "description",
          "dbXrefs",
          "sameAs",
          "status",
          "accessibility",
          "dateCreated",
          "dateModified",
          "datePublished"
        ],
        "title": "SRA"
      },
...
      "Xref": {
        "properties": {
          "identifier": {
            "type": "string",
            "title": "Identifier"
          },
          "type": {
            "type": "string",
            "enum": [
              "biosample",
              "bioproject",
              "sra-experiment",
              "sra-run",
              "sra-sample",
              "sra-study",
              "sra-submission",
              "sra-analysis",
              "jga-study",
              "jga-dataset",
              "jga-dac",
              "jga-policy",
              "gea",
              "geo",
              "hum-id",
              "insdc-assembly",
              "insdc-master",
              "metabobank",
              "pubmed-id",
              "taxonomy",
              "umbrella-bioproject"
            ],
            "title": "Type"
          },
          "url": {
            "type": "string",
            "title": "Url"
          }
        },
        "type": "object",
        "required": [
          "identifier",
          "type",
          "url"
        ],
        "title": "Xref"
      }
    }
  }
}